@@ -17,32 +17,28 @@
     "lint": "ng lint",
     "e2e": "ng e2e",
     "font": "node ./node_modules/webpack/bin/webpack.js",
-    "packagr": "ng-packagr -p ng-package.json",
+    "packagr": "node ./node_modules/ng-packagr/cli/main.js -p ng-package.json",
     "full-package": "npm run packagr && npm pack ./dist && mv *kubernetes-monitor-view-*.tgz kubernetes-monitor-view.tgz"
   },
   "private": true,
-  "peerDependencies": {
-    "@angular/core": ">=^6.0.0"
-  },
+  "peerDependencies": {},
   "dependencies": {},
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.9",
-    "@angular-devkit/build-ng-packagr": "~0.8.0",
+    "@angular-devkit/build-angular": "^0.803.25",
+    "@angular-devkit/build-ng-packagr": "~0.803.25",
     "@angular-devkit/core": "^0.8.9",
     "@angular-devkit/schematics": "^0.8.9",
-    "@angular/animations": "^6.1.10",
-    "@angular/cli": "^6.2.9",
-    "@angular/common": "^6.1.10",
-    "@angular/compiler": "^6.1.10",
-    "@angular/compiler-cli": "^8.1.1",
-    "@angular/core": "^6.1.10",
-    "@angular/forms": "^6.1.10",
-    "@angular/http": "^6.1.10",
-    "@angular/language-service": "^6.1.10",
-    "@angular/platform-browser": "^6.1.10",
-    "@angular/platform-browser-dynamic": "^6.1.10",
-    "@angular/router": "^6.1.10",
-    "@ng-bootstrap/ng-bootstrap": "^2.0.0",
+    "@angular/animations": "^8.2.14",
+    "@angular/cli": "^8.3.25",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "^8.2.14",
+    "@angular/compiler-cli": "^8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/forms": "^8.2.14",
+    "@angular/language-service": "^8.2.14",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "@angular/router": "^8.2.14",
     "@types/bootstrap": "^4.3.1",
     "@types/jasmine": "^2.8.16",
     "@types/jasminewd2": "^2.0.6",
@@ -61,20 +57,20 @@
     "karma-jasmine": "~1.1.1",
     "karma-jasmine-html-reporter": "^0.2.2",
     "moment": "^2.24.0",
-    "ng-packagr": "^5.3.0",
+    "ng-packagr": "^5.7.1",
+    "ngx-bootstrap": "^5.3.2",
     "ngx-logger": "^3.4.5",
-    "ngx-toastr": "^8.6.0",
+    "node-sass": "^4.13.1",
     "octicons": "^8.5.0",
     "popper.js": "^1.15.0",
     "protractor": "^5.4.2",
-    "rxjs": "^6.5.2",
+    "rxjs": "^6.5.4",
     "sass-loader": "^6.0.6",
     "style-loader": "^0.21.0",
     "ts-node": "~3.2.0",
-    "tsickle": "^0.36.0",
     "tslib": "^1.10.0",
     "tslint": "^5.18.0",
-    "typescript": "~3.4.5",
-    "zone.js": "^0.8.29"
+    "typescript": "~3.5.3",
+    "zone.js": "^0.9.1"
   }
 }
